requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic
pydantic-settings
slowapi
python-multipart
gunicorn
